pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pyinstaller>=6.16.0",
    "pymupdf>=1.26.4",
]
